add stress and related phonology
Introduce accent class markers `{R}`, `{D}`, `{F}` for recessive,
dominant, and fixed stress, respectively. All vowels *must* be marked
with one of these in order for correct stress to be inferred.

Unstressed schwas `ы` are now deleted, except in specific exceptional
`CыC`-`CCы́` cases; these must be encoded as `C{ы}Cы{?}` where the final
symbol is an accent class marker used to decide when the final `ы` will be
stressed.

Since the combined lexd-twol transducer produces forms with stress, we
call this the "pregenerator". The actual generator (as with actual
written Abaza) does not mark stress, and is created by composing with a
simple "accent-dropping" second layer of twol:

```drop-accent.twol
Alphabet  ́;

Rules

"drop accent"
́:0 <=> _ ;
```

So the overall composition chain looks like this:

  gloss
  -> morphotactic form w/ AC markers
  -> stressed form w/ phonology effects
  -> unstressed final form

There are many mistakes in the tests which will need to be resolved;
indeed, sources even disagree on what the correct forms are.