Add trailing newline
All files must end with a trailing newline! Otherwise `cat` does not
work properly with them.

Here we attempted to use `cat` to concatenate lexd files; if there is no
trailing newline, the last line of the first file and the first line of
the second file get concatenated, causing syntax errors. :'(